refactor(search): split station_card.dart via part directive
station_card.dart: 479 → 150 LOC. Private widgets (_StatusColumn,
_StationDetails, _StationPriceColumn, _RatingStars, _PriceRow)
moved to station_card_parts.dart as a `part of`, keeping them
private across the library without changing any public API.

No behavior change — all 56 affected tests + full 3734-test suite
pass unchanged. Progress on #563.
